Refactor repository logic and enhance supervisor-intern relationship
Remove the base repository abstraction to simplify repository logic. Introduce Alembic migrations to add a supervisor-intern linking table and additional schema changes, such as handling timestamps with time zones. Refactor the intern and supervisor models to improve relationship handling while enhancing authentication and utilities to support supervisor-specific logic.

+++ src/services/auth_service.py
@@ -5,12 +5,13 @@ from zoneinfo import ZoneInfo
 from fastapi import HTTPException, BackgroundTasks
 from fastapi.params import Depends
 from sqlalchemy.ext.asyncio import AsyncSession
-from starlette.status import HTTP_409_CONFLICT, HTTP_400_BAD_REQUEST, HTTP_401_UNAUTHORIZED
+from starlette.status import HTTP_409_CONFLICT, HTTP_400_BAD_REQUEST, HTTP_401_UNAUTHORIZED, \
+    HTTP_500_INTERNAL_SERVER_ERROR
 
 from src.repositories.general_user_repo import UserRepository
 from src.models.app_models import User, VerificationCode
 from src.db import get_db_session
-from src.repositories.intern_repository import InternRepository
+from src.repositories.intern_repo import InternRepository
 from src.repositories.verification_code_repo import VerificationCodeRepository
 from src.schemas import UserInModel, InternInModel
 from src.schemas.user_schemas import UserOutModel, UserType
@@ -27,6 +28,9 @@ from ..infra.email.contexts import (
 )
 from ..infra.email import send_email
 from ..logger import logger
+from ..repositories.supervisor_repo import SupervisorRepository
+from ..schemas.intern_schemas import InternOutModel
+from ..schemas.supervisor_schemas import SupervisorInModel, SupervisorOutModel
 
 
 class AuthService:
@@ -35,18 +39,20 @@ class AuthService:
         session: Annotated[AsyncSession, Depends(get_db_session)],
         user_repo: Annotated[UserRepository, Depends()],
         intern_repo: Annotated[InternRepository, Depends()],
+        supervisor_repo: Annotated[SupervisorRepository, Depends()],
         code_repo: Annotated[VerificationCodeRepository, Depends()],
         background_task: BackgroundTasks,
     ):
         self.session = session
         self.user_repo = user_repo
         self.intern_repo = intern_repo
+        self.supervisor_repo = supervisor_repo
         self.code_repo = code_repo
         self.background_task = background_task
 
     # TODO: Rate limit
     async def create_unverified_new_user(
-        self, new_user: UserInModel | InternInModel
+        self, new_user: UserInModel | InternInModel | SupervisorInModel
     ) -> dict[str, str]:
         async with self.session.begin():  # Transactional, for atomicity
             existing_user: User = await self.user_repo.get_user_by_email_or_phone(
@@ -75,16 +81,16 @@ class AuthService:
                 # Hash password  when creating a new user
                 new_user.password = hash_password(password=new_user.password)
 
-                unverified_user: User = (
-                    await self.intern_repo.create_new_intern(
+                if isinstance(new_user, InternInModel):
+                    unverified_user: User = await self.intern_repo.create_new_intern(
                         conn=self.session, new_intern=new_user
-                    )  # if the type is Intern
-                    if isinstance(new_user, InternInModel)
-                    else await self.user_repo.create_new_user(  # if type: any other type
-                        conn=self.session, new_user=new_user
                     )
-                    # TODO expand this to accomodate HR (admin) entities
-                )
+                elif isinstance(new_user, SupervisorInModel):
+                    unverified_user: User = await self.supervisor_repo.create_new_supervisor(  # if type: any other type
+                        conn=self.session, new_supervisor=new_user
+                    )
+                else:
+                    unverified_user: User = await self.user_repo.create_new_user(conn=self.session, new_user=new_user)
 
                 code = generate_random_code()
                 await self.code_repo.create_code(
@@ -118,10 +124,18 @@ class AuthService:
                 user_id=verification_code.user_id,
                 values={"verified": True},
             )
+            if not verified_user:
+                raise HTTPException(status_code=HTTP_500_INTERNAL_SERVER_ERROR, detail="Something went wrong with verification")
+
             await self.code_repo.delete_code(conn=self.session, value=code)
 
+            if verified_user.type == UserType.SUPERVISOR:
+                user_to_login = SupervisorOutModel.from_supervisor(verified_user)
+            elif verified_user.type == UserType.INTERN:
+                user_to_login = InternOutModel.from_intern(verified_user)
+
             access_token: str = generate_access_token(
-                UserOutModel.from_user(verified_user),
+                user_to_login=user_to_login,
             )
 
         # Send confirmation mail once user has been created.
@@ -146,9 +160,15 @@ class AuthService:
         if not existing_user.verified:
             raise HTTPException(status_code=HTTP_401_UNAUTHORIZED, detail="Invalid login credentials")
 
-        user_to_login: UserOutModel = UserOutModel.from_user(existing_user)
+        if existing_user.type == UserType.SUPERVISOR:
+            user_to_login: UserOutModel = SupervisorOutModel.from_supervisor(existing_user)
+        elif existing_user.type == UserType.INTERN:
+            user_to_login: UserOutModel = InternOutModel.from_intern(existing_user)
+        else:
+            user_to_login: UserOutModel = UserOutModel.from_user(existing_user)
 
         access_token: str = generate_access_token(user_to_login)
+
         # new_refresh_token: str = await generate_refresh_token(
         #     existing_user.email, self.token_repo
         # )
@@ -254,7 +274,7 @@ class AuthService:
         self.background_task.add_task(
             send_email,
             updated_user.email,
-            context=UpdatedUserContext(values_updated=values_to_update),
+            context=UpdatedUserContext(values_updated=list(values_to_update.keys())),
         )
         return {
             "detail": "Your password has been reset successfully, please proceed to log in"
